# api-contracts/openapi-latest.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User API ",
    "description": "API for user registration and login",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "User Management",
      "description": "Operations related to user registration and login"
    }
  ],
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Register a new user",
        "description": "Creates a new user in the system with a username and password",
        "operationId": "registerUser",
        "requestBody": {
          "description": "User registration payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "User login",
        "description": "Logs in a user and returns a mock authentication token",
        "operationId": "loginUser",
        "requestBody": {
          "description": "User login credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Invalid credentials",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Login successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRegistrationRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "securePass123"
          }
        },
        "description": "User registration request payload"
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username for login",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "description": "Password for login",
            "example": "securePass123"
          }
        },
        "description": "Login request payload"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token",
            "example": "mocked-jwt-token-for-john_doe"
          }
        },
        "description": "Login response containing authentication token"
      }
    }
  }
}